Port TestPlatformAndNodeSuiteScan to serial test package
