+++ tests/e2e/e2e_test.go
@@ -76,220 +76,6 @@ func TestE2E(t *testing.T) {
 		//	},
 		//},
 		testExecution{
-			Name:       "TestPlatformAndNodeSuiteScan",
-			IsParallel: false,
-			TestFn: func(t *testing.T, f *framework.Framework, ctx *framework.TestCtx, namespace string) error {
-				suiteName := "test-suite-two-scans-with-platform"
-
-				workerScanName := fmt.Sprintf("%s-workers-scan", suiteName)
-				selectWorkers := map[string]string{
-					"node-role.kubernetes.io/worker": "",
-				}
-
-				platformScanName := fmt.Sprintf("%s-platform-scan", suiteName)
-
-				exampleComplianceSuite := &compv1alpha1.ComplianceSuite{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      suiteName,
-						Namespace: namespace,
-					},
-					Spec: compv1alpha1.ComplianceSuiteSpec{
-						ComplianceSuiteSettings: compv1alpha1.ComplianceSuiteSettings{
-							AutoApplyRemediations: false,
-						},
-						Scans: []compv1alpha1.ComplianceScanSpecWrapper{
-							{
-								ComplianceScanSpec: compv1alpha1.ComplianceScanSpec{
-									ContentImage: contentImagePath,
-									Profile:      "xccdf_org.ssgproject.content_profile_moderate",
-									Content:      rhcosContentFile,
-									NodeSelector: selectWorkers,
-									ComplianceScanSettings: compv1alpha1.ComplianceScanSettings{
-										Debug: true,
-									},
-								},
-								Name: workerScanName,
-							},
-							{
-								ComplianceScanSpec: compv1alpha1.ComplianceScanSpec{
-									ScanType:     compv1alpha1.ScanTypePlatform,
-									ContentImage: contentImagePath,
-									Profile:      "xccdf_org.ssgproject.content_profile_moderate",
-									Rule:         "xccdf_org.ssgproject.content_rule_ocp_idp_no_htpasswd",
-									Content:      ocpContentFile,
-									ComplianceScanSettings: compv1alpha1.ComplianceScanSettings{
-										Debug: true,
-									},
-								},
-								Name: platformScanName,
-							},
-						},
-					},
-				}
-
-				err := f.Client.Create(goctx.TODO(), exampleComplianceSuite, getCleanupOpts(ctx))
-				if err != nil {
-					return err
-				}
-
-				// Ensure that all the scans in the suite have finished and are marked as Done
-				err = waitForSuiteScansStatus(t, f, namespace, suiteName, compv1alpha1.PhaseDone,
-					compv1alpha1.ResultNonCompliant)
-				if err != nil {
-					return err
-				}
-
-				// At this point, both scans should be non-compliant given our current content
-				err = scanResultIsExpected(t, f, namespace, workerScanName, compv1alpha1.ResultNonCompliant)
-				if err != nil {
-					return err
-				}
-
-				// The profile should find one check for an htpasswd IDP, so we should be compliant.
-				err = scanResultIsExpected(t, f, namespace, platformScanName, compv1alpha1.ResultCompliant)
-				if err != nil {
-					return err
-				}
-
-				// Each scan should produce two remediations
-				workerRemediations := []string{
-					fmt.Sprintf("%s-no-empty-passwords", workerScanName),
-					fmt.Sprintf("%s-no-direct-root-logins", workerScanName),
-				}
-				err = assertHasRemediations(t, f, suiteName, workerScanName, "worker", workerRemediations)
-				if err != nil {
-					return err
-				}
-
-				// TODO: Add check for future API remediation
-				//platformRemediations := []string{
-				//	fmt.Sprintf("%s-no-empty-passwords", platformScanName),
-				//	fmt.Sprintf("%s-no-direct-root-logins", platformScanName),
-				//}
-				//err = assertHasRemediations(t, f, suiteName, platformScanName, "master", platformRemediations)
-				//if err != nil {
-				//	return err
-				//}
-
-				// Test a fail result from the platform scan. This fails the HTPasswd IDP check.
-				if _, err := f.KubeClient.CoreV1().Secrets("openshift-config").Create(goctx.TODO(), &corev1.Secret{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "htpass",
-						Namespace: "openshift-config",
-					},
-					Type: corev1.SecretTypeOpaque,
-					Data: map[string][]byte{
-						"htpasswd": []byte("bob:$2y$05$OyjQO7M2so4hRJW0aS9yie9KJ0wXv80XFWyEsApUZFURqE37aVR/a"),
-					},
-				}, metav1.CreateOptions{}); err != nil {
-					return err
-				}
-
-				defer func() {
-					err := f.KubeClient.CoreV1().Secrets("openshift-config").Delete(goctx.TODO(), "htpass", metav1.DeleteOptions{})
-					if err != nil {
-						E2ELogf(t, "could not clean up openshift-config/htpass test secret: %v", err)
-					}
-				}()
-
-				fetchedOauth := &configv1.OAuth{}
-				err = f.Client.Get(goctx.TODO(), types.NamespacedName{Name: "cluster"}, fetchedOauth)
-				if err != nil {
-					return err
-				}
-
-				oauthUpdate := fetchedOauth.DeepCopy()
-				oauthUpdate.Spec = configv1.OAuthSpec{
-					IdentityProviders: []configv1.IdentityProvider{
-						{
-							Name:          "my_htpasswd_provider",
-							MappingMethod: "claim",
-							IdentityProviderConfig: configv1.IdentityProviderConfig{
-								Type: "HTPasswd",
-								HTPasswd: &configv1.HTPasswdIdentityProvider{
-									FileData: configv1.SecretNameReference{
-										Name: "htpass",
-									},
-								},
-							},
-						},
-					},
-				}
-
-				err = f.Client.Update(goctx.TODO(), oauthUpdate)
-				if err != nil {
-					E2ELogf(t, "error updating idp: %v", err)
-					return err
-				}
-
-				defer func() {
-					fetchedOauth := &configv1.OAuth{}
-					err := f.Client.Get(goctx.TODO(), types.NamespacedName{Name: "cluster"}, fetchedOauth)
-					if err != nil {
-						E2ELogf(t, "error restoring idp: %v", err)
-					} else {
-						oauth := fetchedOauth.DeepCopy()
-						// Make sure it's cleared out
-						oauth.Spec = configv1.OAuthSpec{
-							IdentityProviders: nil,
-						}
-						err = f.Client.Update(goctx.TODO(), oauth)
-						if err != nil {
-							E2ELogf(t, "error restoring idp: %v", err)
-						}
-					}
-				}()
-
-				suiteName = "test-suite-two-scans-with-platform-2"
-				platformScanName = fmt.Sprintf("%s-platform-scan-2", suiteName)
-				exampleComplianceSuite = &compv1alpha1.ComplianceSuite{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      suiteName,
-						Namespace: namespace,
-					},
-					Spec: compv1alpha1.ComplianceSuiteSpec{
-						ComplianceSuiteSettings: compv1alpha1.ComplianceSuiteSettings{
-							AutoApplyRemediations: false,
-						},
-						Scans: []compv1alpha1.ComplianceScanSpecWrapper{
-							{
-								ComplianceScanSpec: compv1alpha1.ComplianceScanSpec{
-									ScanType:     compv1alpha1.ScanTypePlatform,
-									ContentImage: contentImagePath,
-									Profile:      "xccdf_org.ssgproject.content_profile_moderate",
-									Rule:         "xccdf_org.ssgproject.content_rule_ocp_idp_no_htpasswd",
-									Content:      ocpContentFile,
-									ComplianceScanSettings: compv1alpha1.ComplianceScanSettings{
-										Debug: true,
-									},
-								},
-								Name: platformScanName,
-							},
-						},
-					},
-				}
-
-				err = f.Client.Create(goctx.TODO(), exampleComplianceSuite, getCleanupOpts(ctx))
-				if err != nil {
-					return err
-				}
-
-				// Ensure that all the scans in the suite have finished and are marked as Done
-				err = waitForSuiteScansStatus(t, f, namespace, suiteName, compv1alpha1.PhaseDone,
-					compv1alpha1.ResultNonCompliant)
-				if err != nil {
-					return err
-				}
-
-				err = scanResultIsExpected(t, f, namespace, platformScanName, compv1alpha1.ResultNonCompliant)
-				if err != nil {
-					return err
-				}
-
-				return nil
-			},
-		},
-		testExecution{
 			Name:       "TestUpdateRemediation",
 			IsParallel: false,
 			TestFn: func(t *testing.T, f *framework.Framework, ctx *framework.Context, namespace string) error {
